pyproject.toml
[project]
name = "cib-chatbot-serverside"
version = "0.1.0"
description = "RAG-based chatbot API using FastAPI, LangChain, and PostgreSQL with pgvector"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "langchain (==0.2.2)",
    "langchain-ollama (==0.1.3)",
    "langchain-community (==0.2.3)",
    "langchain-chroma (==0.1.2)",
    "langchain-core (>=0.2.0,<0.3.0)",
    "langchain-text-splitters (==0.2.2)",
    "langchain-postgres (>=0.0.17,<0.0.18)",
    "unstructured (==0.14.4)",
    "pypdf (>=6.7.0,<7.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pgvector (>=0.3.0,<0.4.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (==0.40.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[project.scripts]
chatbot-server = "cib_chatbot_serverside.main:main"
sync-docs = "cib_chatbot_serverside.scripts.sync_documents:main"

[tool.poetry]
packages = [{include = "cib_chatbot_serverside", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
